Tidy up gradle plugins
