+++ appAndroid/appAndroid/build.gradle.kts
@@ -1,5 +1,5 @@
 plugins {
-    id("android-application-plugin")
+    id("android-application-binary-plugin")
 }
 
 android {
