+++ appAndroid/common-test/build.gradle.kts
@@ -1,7 +1,7 @@
 import com.gchristov.newsfeed.gradleplugins.Deps
 
 plugins {
-    id("android-library-plugin")
+    id("android-base-plugin") // Avoids circular dependencies in android-module-plugin
 }
 
 /*
@@ -10,5 +10,5 @@ source sets, and marked as `api`, rather than 'test'. This is because 'test' sou
 dependencies and code are local to the relevant module and cannot be accesses by other modules.
  */
 dependencies {
-    api(Deps.Tests.junit)
+    api(Deps.Android.Tests.junit)
 }
