+++ gradle-plugins/settings.gradle.kts
