+++ gradle-plugins/src/main/kotlin/com/gchristov/newsfeed/gradleplugins/android/AndroidBinaryPlugin.kt
@@ -1,0 +1,13 @@
+package com.gchristov.newsfeed.gradleplugins.android
+
+import org.gradle.api.Plugin
+import org.gradle.api.Project
+
+class AndroidApplicationBinaryPlugin : Plugin<Project> {
+    override fun apply(target: Project) {
+        target.run {
+            plugins.apply("com.android.application")
+            plugins.apply("android-base-plugin")
+        }
+    }
+}
