+++ gradle-plugins/src/main/kotlin/com/gchristov/newsfeed/gradleplugins/multiplatform/MplBasePlugin.kt
@@ -1,0 +1,50 @@
+package com.gchristov.newsfeed.gradleplugins.multiplatform
+
+import com.android.build.gradle.BaseExtension
+import com.gchristov.newsfeed.gradleplugins.Deps
+import org.gradle.api.Plugin
+import org.gradle.api.Project
+import org.gradle.kotlin.dsl.configure
+import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
+import org.jetbrains.kotlin.gradle.plugin.mpp.KotlinNativeTarget
+
+class MplBasePlugin : Plugin<Project> {
+    override fun apply(target: Project) {
+        target.run {
+            plugins.apply("com.android.library")
+            plugins.apply("org.jetbrains.kotlin.multiplatform")
+            extensions.configure<BaseExtension> {
+                compileSdkVersion(Deps.Android.compileSdk)
+                defaultConfig {
+                    minSdk = Deps.Android.minSdk
+                    targetSdk = Deps.Android.targetSdk
+                    testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
+                }
+                // Include AndroidManifest.xml, if it exists.
+                sourceSets {
+                    maybeCreate("main").manifest {
+                        srcFile("src/androidMain/AndroidManifest.xml")
+                    }
+                }
+                // Resolves "N files found with path 'META-INF/XXX'" errors
+                packagingOptions {
+                    resources.excludes.add("META-INF/AL2.0")
+                    resources.excludes.add("META-INF/LGPL2.1")
+                }
+            }
+            extensions.configure<KotlinMultiplatformExtension> {
+                android()
+                val iosTarget: (String, KotlinNativeTarget.() -> Unit) -> KotlinNativeTarget = when {
+                    System.getenv("SDK_NAME")?.startsWith("iphoneos") == true -> ::iosArm64
+                    System.getenv("NATIVE_ARCH")?.startsWith("arm") == true -> ::iosSimulatorArm64
+                    else -> ::iosX64
+                }
+                iosTarget("ios") {}
+                // Allows libraries to not have to specify all source sets unless required.
+                sourceSets.maybeCreate("commonMain")
+                sourceSets.maybeCreate("androidMain")
+                sourceSets.maybeCreate("iosMain")
+            }
+        }
+    }
+}
