+++ kmm/kmm-common-kotlin/build.gradle.kts
@@ -1,20 +1,20 @@
-import com.gchristov.newsfeed.kmmgradleplugins.Deps
+import com.gchristov.newsfeed.gradleplugins.Deps
 
 plugins {
-    id("kmm-platform-plugin")
+    id("mpl-base-plugin")
 }
 
 kotlin {
     sourceSets {
         val commonMain by getting {
             dependencies {
-                api(Deps.Kotlin.coroutinesCore)
-                api(Deps.Kotlin.dateTime)
+                api(Deps.Multiplatform.Kotlin.coroutinesCore)
+                api(Deps.Multiplatform.Kotlin.dateTime)
             }
         }
         val androidMain by getting {
             dependencies {
-                api(Deps.Kotlin.coroutinesAndroid)
+                api(Deps.Multiplatform.Kotlin.coroutinesAndroid)
             }
         }
     }
