+++ kmm/kmm-common-persistence/build.gradle.kts
@@ -1,25 +1,25 @@
-import com.gchristov.newsfeed.kmmgradleplugins.Deps
+import com.gchristov.newsfeed.gradleplugins.Deps
 
 plugins {
-    id("kmm-module-plugin")
+    id("mpl-module-plugin")
 }
 
 kotlin {
     sourceSets {
         val commonMain by getting {
             dependencies {
-                implementation(Deps.SqlDelight.runtime)
-                api(Deps.SharedPreferences.multiplatformSettings)
+                implementation(Deps.Multiplatform.SqlDelight.runtime)
+                api(Deps.Multiplatform.SharedPreferences.multiplatformSettings)
             }
         }
         val androidMain by getting {
             dependencies {
-                implementation(Deps.SqlDelight.driverAndroid)
+                implementation(Deps.Multiplatform.SqlDelight.driverAndroid)
             }
         }
         val iosMain by getting {
             dependencies {
-                implementation(Deps.SqlDelight.driverNative)
+                implementation(Deps.Multiplatform.SqlDelight.driverNative)
             }
         }
     }
