--- kmm/kmm-gradle-plugins/settings.gradle.kts
@@ -1,8 +1,0 @@
-pluginManagement {
-    repositories {
-        google()
-        gradlePluginPortal()
-        mavenCentral()
-    }
-}
-rootProject.name = "kmm-gradle-plugins"
