+++ kmm/kmm-post-test-fixtures/build.gradle.kts
@@ -1,5 +1,5 @@
 plugins {
-    id("kmm-module-plugin")
+    id("mpl-module-plugin")
 }
 
 kotlin {
