+++ kmm/kmm-umbrella/build.gradle.kts
@@ -1,9 +1,9 @@
-import com.gchristov.newsfeed.kmmgradleplugins.Deps
+import com.gchristov.newsfeed.gradleplugins.Deps
 import org.jetbrains.kotlin.gradle.plugin.mpp.KotlinNativeTarget
 
 plugins {
-    // Normally, this would be kmm-module-plugin but since this is an umbrella module this is enough
-    id("kmm-platform-plugin")
+    // Normally, this would be mpl-module-plugin but since this is an umbrella module this is enough
+    id("mpl-base-plugin")
 }
 
 kotlin {
@@ -13,7 +13,7 @@ kotlin {
      https://touchlab.co/multiple-kotlin-frameworks-in-application/
      */
     val exportedDependencies = listOf(
-        Deps.Kotlin.coroutinesCore, // Needed for coroutine dispatchers
+        Deps.Multiplatform.Kotlin.coroutinesCore, // Needed for coroutine dispatchers
         projects.kmmCommonTest,
         projects.kmmFeed,
         projects.kmmFeedData,
